requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.30.0

# API and web framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0

# Data validation and models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client and utilities
httpx>=0.24.0
python-dotenv>=1.0.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data processing and Excel export
pandas>=2.0.0
openpyxl>=3.0.0
numpy>=1.24.0

# Stock market data
yfinance>=0.2.0

# PDF processing
PyMuPDF>=1.23.0
camelot-py[cv]>=0.11.0

# PDF generation
reportlab>=4.0.0

# Templating
jinja2>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development tools
black>=23.0.0
ruff>=0.0.285
mypy>=1.5.0

